Model Serializer Map
Hi,

I was using your library for a project and needed the {model:serializer} mapping to be configured in the websites' settings. The `dict` provided in this commit allows one to configure this in the sites settings and access the `Serializers` lazily; since importing serializers within the settings module occurs before all applications are loaded.  

Usage example :

    PACKAGE_MODEL_FIELD_SERIALIZERS = ModelSerializerMap(**{"APPLICATION.models.MODEL":"APPLICATION.serializers.SERIALIZER"})

Please consider adding this to the libray.